tests(tui): generalize test harness
