+++ crates/but/src/tui/mod.rs
@@ -7,6 +7,8 @@ pub use table::types::Table;
 
 pub mod text;
 
+pub mod event_polling;
+
 pub mod get_text;
 
 pub(crate) mod diff_viewer;
@@ -27,7 +29,16 @@ use crossterm::{
 use ratatui::{
     Terminal, TerminalOptions, Viewport,
     backend::{CrosstermBackend, TestBackend},
+    prelude::Backend,
 };
+
+use crate::{
+    tui::event_polling::EventPolling,
+    utils::{InputOutputChannel, WriteWithUtils},
+};
+
+#[cfg(test)]
+pub mod test_utils;
 
 type PanicHook = Box<dyn Fn(&std::panic::PanicHookInfo<'_>) + Send + Sync>;
 
@@ -253,5 +264,45 @@ impl Drop for SuspendGuard<'_> {
         if self.0.config.uses_alt_screen {
             _ = self.0.terminal.clear();
         }
+    }
+}
+
+pub trait Tui {
+    type UpdateContext<'a>;
+
+    fn update<T, E>(
+        &mut self,
+        terminal_guard: &mut T,
+        event_polling: E,
+        events: &mut Vec<crossterm::event::Event>,
+        out: &mut dyn TuiInputOutputChannel,
+        update_ctx: &mut Self::UpdateContext<'_>,
+    ) -> anyhow::Result<()>
+    where
+        T: TerminalGuard,
+        anyhow::Error: From<<T::Backend as Backend>::Error>,
+        E: EventPolling;
+
+    fn render<T>(&mut self, terminal_guard: &mut T) -> anyhow::Result<()>
+    where
+        T: TerminalGuard,
+        anyhow::Error: From<<T::Backend as Backend>::Error>;
+}
+
+/// Required to abstract over input/output channels for the TUI.
+///
+/// In production we want to require `InputOutputChannel`. This means the caller must check that
+/// input is actually supported and return an error otherwise. However in tests we don't want to
+/// enforce that.
+///
+/// So this trait exists such that we can make a fake to use in tests that panics on
+/// `prompt_single_line`.
+pub trait TuiInputOutputChannel: WriteWithUtils {
+    fn prompt_single_line(&mut self, prompt: &str) -> anyhow::Result<Option<String>>;
+}
+
+impl TuiInputOutputChannel for InputOutputChannel<'_> {
+    fn prompt_single_line(&mut self, prompt: &str) -> anyhow::Result<Option<String>> {
+        InputOutputChannel::prompt_single_line(self, prompt)
     }
 }
